Ignoring Content-Transfer-Encoding headers with missing values.
A Content-Transfer-Encoding header with a missing value
would be handled the same as an unknown encoding value:
a message would be included in the body of the converted message
stating the encoding was unknown as well as the unknown encoding
value. However, if the value was missing, hypermail would use
the value of an unitialized variable.