"base64 decode assumes input lines are in multiples of 4 #96"
Convert base64 implementation of stateless to stateful, so you can
feed in successive lines and get back the original binary.  The
previous stateless implementation assumes that the input has
no residual after processing a single line, which is not a semantic
guaranteed by the base64-in-email standards (and, in fact, was
found while processing exports from Tutanota).