New functions to create a tree of message body part nodes, created while parsing a message. These include functions to dump the tree and to flatten it to the body structure used after parsing.
Adding hm_stack, a generic stack, and moving some of our ad-hoc
stacks to it (missing boundary).

Deprecated with #ifdef CHARSETSP code that previously used to
memorize the charsets of body parts in order to decide which was
the prefered one for rendering the message. The message nodes allow
for better heuristics. That deprecated code will be removed in a
commit further down the road.

Removed errors and warnings detected by gcc -Wall