Fixes issues related to parsing invalid headers in mail messages. Improves the handling of UTF-8 header values.
Fixes an issue when parsing a message with one or more invalid email
headers could make hypermail crash. If hypermail detects an invalid
header, it will mark it as such and skip it during the rest of the
parsing and markup generation. Some examples of invalid headers
are those missing a header name, value, ':\s', etc.

extend mdecode_RFC2047() so that;
1) it partially supports RFC6532
2) if a header value is not valid UTF-8,
replaces it with "(invalid string)"
3) if a header value is given as a binary unencoded string,
it tries to detect the charset with libchardet. In case of failure,
it will replace the header value with "(invalid string)"