Filter the charset attribute value to remove "DEFAULT_CHARSET" and "_CHARSET"
Some old messages or broken UA used "DEFAULT_CHARSET" or "foo_CHARSET"
as a value for the charset attribute. We now filter out those broken
values from the charset value. In the case of "DEFAULT_CHARSET", the
end result will be equivalent to not having a charset attribute.  In
that case, the parser will either use the charset associated with
another header value or the first body part or, by default, use
US-ASCII.